Add unit test for invert binary tree.

+++ data_structures/source/test/trees/binary_trees/test_in_order_traversal.cpp
@@ -11,7 +11,7 @@ TEST_F(SetupBinaryTree, test_inorder_traversal)
 {
 	auto root = SetupBinaryTree::binary_tree_root();
 	auto result = in_order_traversal(root);
-	for(int index =0; index < inorder.size(); ++index)
+	for(size_t index =0; index < inorder.size(); ++index)
 	{
 		EXPECT_EQ(result[index], inorder[index]);
 	}
@@ -21,7 +21,7 @@ TEST_F(SetupBinaryTree, test_inorder_traversal_recursive)
 {
 	auto root = SetupBinaryTree::binary_tree_root();
 	auto result = in_order_traversal_recursive(root);
-	for(int index =0; index < inorder.size(); ++index)
+	for(size_t index =0; index < inorder.size(); ++index)
 	{
 		EXPECT_EQ(result[index], inorder[index]);
 	}
